carga primer nivel

+++ backend/mipBackend/Dtos/InmuebleDtos/InmuebleRequestDto.cs
@@ -7,8 +7,9 @@
 
         public string? Direccion { get; set; }
 
-        public decimal Precio { get; set; }
+        public int Precio { get; set; }
 
         public string? Picture { get; set; }
+
     }
 }
